+++ dtool_create/dataset.py
@@ -298,6 +298,13 @@ def edit(dataset_uri):
             config_path=CONFIG_PATH
         )
     readme_content = dataset.get_readme_content()
+
+    try:
+        # Python2 compatibility.
+        readme_content = unicode(readme_content, "utf-8")
+    except NameError:
+        pass
+
     edited_content = click.edit(readme_content)
     if edited_content is not None:
         _validate_and_put_readme(dataset, edited_content)